requirements.txt
asgiref==3.8.1
Django==4.2.16
django-browser-reload==1.15.0
django-taggit==6.1.0
django-tinymce==4.1.0
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
